pyproject.toml
[tool.poetry]
name = "drf-advanced-parwiz"
version = "0.1.0"
description = ""
authors = ["TestardR <romain.rtestard@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.9"
Django = "^3.1.4"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest-django = "^4.1.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
